Moved Zemismart sensor quirk to Tuya power meters
